Fix newsletter related broken eslint rules
