+++ assets/js/src/newsletters/newsletters.jsx
@@ -4,6 +4,7 @@ import { Router, Route, IndexRedirect, useRouterHistory } from 'react-router';
 import { createHashHistory } from 'history';
 import Hooks from 'wp-js-hooks';
 import _ from 'underscore';
+import PropTypes from 'prop-types';
 
 import NewsletterTypes from 'newsletters/types.jsx';
 import NewsletterTemplates from 'newsletters/templates.jsx';
@@ -24,6 +25,10 @@ class App extends React.Component {
     return this.props.children;
   }
 }
+
+App.propTypes = {
+  children: PropTypes.element.isRequired,
+};
 
 const container = document.getElementById('newsletters_container');
