+++ assets/js/src/newsletters/send/standard.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
-import jQuery from 'jquery';
 import _ from 'underscore';
 import MailPoet from 'mailpoet';
 import Hooks from 'wp-js-hooks';
+import PropTypes from 'prop-types';
+
+import DateTime from 'newsletters/send/date_time.jsx';
 
 const currentTime = window.mailpoet_current_time || '00:00';
 const defaultDateTime = `${window.mailpoet_current_date} 00:00:00`;
@@ -10,257 +12,30 @@ const timeOfDayItems = window.mailpoet_schedule_time_of_day;
 const dateDisplayFormat = window.mailpoet_date_display_format;
 const dateStorageFormat = window.mailpoet_date_storage_format;
 
-const datepickerTranslations = {
-  closeText: MailPoet.I18n.t('close'),
-  currentText: MailPoet.I18n.t('today'),
-  nextText: MailPoet.I18n.t('next'),
-  prevText: MailPoet.I18n.t('previous'),
-  monthNames: [
-    MailPoet.I18n.t('january'),
-    MailPoet.I18n.t('february'),
-    MailPoet.I18n.t('march'),
-    MailPoet.I18n.t('april'),
-    MailPoet.I18n.t('may'),
-    MailPoet.I18n.t('june'),
-    MailPoet.I18n.t('july'),
-    MailPoet.I18n.t('august'),
-    MailPoet.I18n.t('september'),
-    MailPoet.I18n.t('october'),
-    MailPoet.I18n.t('november'),
-    MailPoet.I18n.t('december'),
-  ],
-  monthNamesShort: [
-    MailPoet.I18n.t('januaryShort'),
-    MailPoet.I18n.t('februaryShort'),
-    MailPoet.I18n.t('marchShort'),
-    MailPoet.I18n.t('aprilShort'),
-    MailPoet.I18n.t('mayShort'),
-    MailPoet.I18n.t('juneShort'),
-    MailPoet.I18n.t('julyShort'),
-    MailPoet.I18n.t('augustShort'),
-    MailPoet.I18n.t('septemberShort'),
-    MailPoet.I18n.t('octoberShort'),
-    MailPoet.I18n.t('novemberShort'),
-    MailPoet.I18n.t('decemberShort'),
-  ],
-  dayNames: [
-    MailPoet.I18n.t('sunday'),
-    MailPoet.I18n.t('monday'),
-    MailPoet.I18n.t('tuesday'),
-    MailPoet.I18n.t('wednesday'),
-    MailPoet.I18n.t('thursday'),
-    MailPoet.I18n.t('friday'),
-    MailPoet.I18n.t('saturday'),
-  ],
-  dayNamesShort: [
-    MailPoet.I18n.t('sundayShort'),
-    MailPoet.I18n.t('mondayShort'),
-    MailPoet.I18n.t('tuesdayShort'),
-    MailPoet.I18n.t('wednesdayShort'),
-    MailPoet.I18n.t('thursdayShort'),
-    MailPoet.I18n.t('fridayShort'),
-    MailPoet.I18n.t('saturdayShort'),
-  ],
-  dayNamesMin: [
-    MailPoet.I18n.t('sundayMin'),
-    MailPoet.I18n.t('mondayMin'),
-    MailPoet.I18n.t('tuesdayMin'),
-    MailPoet.I18n.t('wednesdayMin'),
-    MailPoet.I18n.t('thursdayMin'),
-    MailPoet.I18n.t('fridayMin'),
-    MailPoet.I18n.t('saturdayMin'),
-  ],
-};
-
-class DateText extends React.Component {
-  onChange = (event) => {
-    const changeEvent = event;
-    // Swap display format to storage format
-    const displayDate = changeEvent.target.value;
-    const storageDate = this.getStorageDate(displayDate);
-
-    changeEvent.target.value = storageDate;
-    this.props.onChange(changeEvent);
-  };
-
-  componentDidMount() {
-    const $element = jQuery(this.dateInput);
-    const that = this;
-    if ($element.datepicker) {
-      // Override jQuery UI datepicker Date parsing and formatting
-      jQuery.datepicker.parseDate = function parseDate(format, value) {
-        // Transform string format to Date object
-        return MailPoet.Date.toDate(value, {
-          parseFormat: dateDisplayFormat,
-          format,
-        });
-      };
-      jQuery.datepicker.formatDate = function formatDate(format, value) {
-        // Transform Date object to string format
-        const newValue = MailPoet.Date.format(value, {
-          format,
-        });
-        return newValue;
-      };
-
-      $element.datepicker(_.extend({
-        dateFormat: this.props.displayFormat,
-        isRTL: false,
-        onSelect: function onSelect(value) {
-          that.onChange({
-            target: {
-              name: that.getFieldName(),
-              value,
-            },
-          });
-        },
-      }, datepickerTranslations));
-
-      this.datepickerInitialized = true;
-    }
-  }
-
-  componentWillUnmount() {
-    if (this.datepickerInitialized) {
-      jQuery(this.dateInput).datepicker('destroy');
-    }
-  }
-
-  getFieldName = () => {
-    return this.props.name || 'date';
-  };
-
-  getDisplayDate = (date) => {
-    return MailPoet.Date.format(date, {
-      parseFormat: this.props.storageFormat,
-      format: this.props.displayFormat,
-    });
-  };
-
-  getStorageDate = (date) => {
-    return MailPoet.Date.format(date, {
-      parseFormat: this.props.displayFormat,
-      format: this.props.storageFormat,
-    });
-  };
-
-  render() {
-    return (
-      <input
-        type="text"
-        size="30"
-        name={this.getFieldName()}
-        value={this.getDisplayDate(this.props.value)}
-        readOnly
-        disabled={this.props.disabled}
-        onChange={this.onChange}
-        ref={(c) => { this.dateInput = c; }}
-        {...this.props.validation}
-      />
-    );
-  }
-}
-
-class TimeSelect extends React.Component {
-  render() {
-    const options = Object.keys(timeOfDayItems).map(
-      value => (
-        <option
-          key={`option-${timeOfDayItems[value]}`}
-          value={value}
-        >
-          { timeOfDayItems[value] }
-        </option>
-      )
-    );
-
-    return (
-      <select
-        name={this.props.name || 'time'}
-        value={this.props.value}
-        disabled={this.props.disabled}
-        onChange={this.props.onChange}
-        {...this.props.validation}
-      >
-        {options}
-      </select>
-    );
-  }
-}
-
-class DateTime extends React.Component {
-  DATE_TIME_SEPARATOR = ' ';
-
-  componentWillReceiveProps(nextProps) {
-    this.setState(this.buildStateFromProps(nextProps));
-  }
-
-  buildStateFromProps = (props) => {
-    const value = props.value || defaultDateTime;
-    const [date, time] = value.split(this.DATE_TIME_SEPARATOR);
-    return {
-      date,
-      time,
-    };
-  };
-
-  handleChange = (event) => {
-    const newState = {};
-    newState[event.target.name] = event.target.value;
-
-    this.setState(newState, this.propagateChange);
-  };
-
-  propagateChange = () => {
-    if (this.props.onChange) {
-      this.props.onChange({
-        target: {
-          name: this.props.name || '',
-          value: this.getDateTime(),
-        },
-      });
-    }
-  };
-
-  getDateTime = () => {
-    return [this.state.date, this.state.time].join(this.DATE_TIME_SEPARATOR);
-  };
-
-  state = this.buildStateFromProps(this.props);
-
-  render() {
-    return (
-      <span>
-        <DateText
-          name="date"
-          value={this.state.date}
-          onChange={this.handleChange}
-          displayFormat={dateDisplayFormat}
-          storageFormat={dateStorageFormat}
-          disabled={this.props.disabled}
-          validation={this.props.dateValidation}
-        />
-        <TimeSelect
-          name="time"
-          value={this.state.time}
-          onChange={this.handleChange}
-          disabled={this.props.disabled}
-          validation={this.props.timeValidation}
-        />
-      </span>
-    );
-  }
-}
-
 class StandardScheduling extends React.Component {
   getCurrentValue = () => {
+    const schedulingOptions = {
+      isScheduled: '0',
+      scheduledAt: defaultDateTime,
+    };
     return _.defaults(
       this.props.item[this.props.field.name] || {},
-      {
-        isScheduled: '0',
-        scheduledAt: defaultDateTime,
-      }
+      schedulingOptions
     );
+  };
+
+  getDateValidation = () => ({
+    'data-parsley-required': true,
+    'data-parsley-required-message': MailPoet.I18n.t('noScheduledDateError'),
+    'data-parsley-errors-container': '#mailpoet_scheduling',
+  });
+
+  isScheduled = () => this.getCurrentValue().isScheduled === '1';
+
+  handleCheckboxChange = (event) => {
+    const changeEvent = event;
+    changeEvent.target.value = this.isScheduledInput.checked ? '1' : '0';
+    return this.handleValueChange(changeEvent);
   };
 
   handleValueChange = (event) => {
@@ -276,24 +51,6 @@ class StandardScheduling extends React.Component {
     });
   };
 
-  handleCheckboxChange = (event) => {
-    const changeEvent = event;
-    changeEvent.target.value = this.isScheduledInput.checked ? '1' : '0';
-    return this.handleValueChange(changeEvent);
-  };
-
-  isScheduled = () => {
-    return this.getCurrentValue().isScheduled === '1';
-  };
-
-  getDateValidation = () => {
-    return {
-      'data-parsley-required': true,
-      'data-parsley-required-message': MailPoet.I18n.t('noScheduledDateError'),
-      'data-parsley-errors-container': '#mailpoet_scheduling',
-    };
-  };
-
   render() {
     let schedulingOptions;
 
@@ -306,6 +63,10 @@ class StandardScheduling extends React.Component {
             onChange={this.handleValueChange}
             disabled={this.props.field.disabled}
             dateValidation={this.getDateValidation()}
+            defaultDateTime={defaultDateTime}
+            timeOfDayItems={timeOfDayItems}
+            dateDisplayFormat={dateDisplayFormat}
+            dateStorageFormat={dateStorageFormat}
           />
           &nbsp;
           <span>
@@ -331,6 +92,19 @@ class StandardScheduling extends React.Component {
     );
   }
 }
+
+StandardScheduling.propTypes = {
+  item: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  field: PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    disabled: PropTypes.bool,
+  }).isRequired,
+  onValueChange: PropTypes.func.isRequired,
+};
+
+StandardScheduling.defaultProps = {
+  item: {},
+};
 
 let fields = [
   {
