+++ assets/js/src/newsletters/types/notification/scheduling.jsx
@@ -1,5 +1,6 @@
 import _ from 'underscore';
 import React from 'react';
+import PropTypes from 'prop-types';
 import Select from 'form/fields/select.jsx';
 import {
   intervalValues,
@@ -35,9 +36,7 @@ const nthWeekDayField = {
 };
 
 class NotificationScheduling extends React.Component {
-  getCurrentValue = () => {
-    return (this.props.item[this.props.field.name] || {});
-  };
+  getCurrentValue = () => this.props.item[this.props.field.name] || {};
 
   handleValueChange = (name, value) => {
     const oldValue = this.getCurrentValue();
@@ -53,40 +52,11 @@ class NotificationScheduling extends React.Component {
     });
   };
 
-  handleIntervalChange = (event) => {
-    return this.handleValueChange(
-      'intervalType',
-      event.target.value
-    );
-  };
-
-  handleTimeOfDayChange = (event) => {
-    return this.handleValueChange(
-      'timeOfDay',
-      event.target.value
-    );
-  };
-
-  handleWeekDayChange = (event) => {
-    return this.handleValueChange(
-      'weekDay',
-      event.target.value
-    );
-  };
-
-  handleMonthDayChange = (event) => {
-    return this.handleValueChange(
-      'monthDay',
-      event.target.value
-    );
-  };
-
-  handleNthWeekDayChange = (event) => {
-    return this.handleValueChange(
-      'nthWeekDay',
-      event.target.value
-    );
-  };
+  handleIntervalChange = event => this.handleValueChange('intervalType', event.target.value);
+  handleTimeOfDayChange = event => this.handleValueChange('timeOfDay', event.target.value);
+  handleWeekDayChange = event => this.handleValueChange('weekDay', event.target.value);
+  handleMonthDayChange = event => this.handleValueChange('monthDay', event.target.value);
+  handleNthWeekDayChange = event => this.handleValueChange('nthWeekDay', event.target.value);
 
   render() {
     const value = this.getCurrentValue();
@@ -152,5 +122,13 @@ class NotificationScheduling extends React.Component {
     );
   }
 }
+
+NotificationScheduling.propTypes = {
+  item: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
+  field: PropTypes.shape({
+    name: PropTypes.string,
+  }).isRequired,
+  onValueChange: PropTypes.func.isRequired,
+};
 
 module.exports = NotificationScheduling;
